requirements.txt
python-dotenv
httpx
requests
openai
